refactor fron detailactivity to detailfragment
