+++ app/src/main/res/layout/activity_detail.xml
@@ -1,5 +1,5 @@
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" android:id="@+id/fragment"
+    xmlns:tools="http://schemas.android.com/tools" android:id="@+id/fragment_detail"
     android:name="com.example.mkoldobsky.popmovies.DetailFragment"
     tools:layout="@layout/fragment_detail" android:layout_width="match_parent"
     android:layout_height="match_parent" />
